requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
pymysql
python-dotenv
cryptography
